Rename IngestJobStatusTestData
